+++ packages/web3/src/utils/index.ts
@@ -1,0 +1,15 @@
+export const getPlatform = () => {
+  const userAgent = navigator.userAgent.toLowerCase();
+
+  if (userAgent.includes('chrome')) {
+    return 'Chrome';
+  } else if (userAgent.includes('firefox')) {
+    return 'Firefox';
+  } else if (userAgent.includes('edge')) {
+    return 'Edge';
+  } else if (userAgent.includes('safari')) {
+    return 'Safari';
+  } else {
+    return 'Other';
+  }
+};
